build: Don't just check an existence for CORES_FILE
CORES_FILE returns the result of "cat minibuild/cores", which returns a number. This simply broke the "if [ -e ]" switch, which checks for a file (and instead receives a number instead).

Fix this by making CORES_FILE a static reference to the file and then use cat to check its value when it's found.

* Also indents stuff properly