build: Properly use "if..elif..else"
We used "if..else" before, this causes the script to fallback to "build"
if the command inputted is invalid. Fix this by using elif, so that we
can add an error message for invalid commands.

Also add "config" function as a shortcut to launch installation script
again. Might want to revamp this later.