build & install: Fix "config" behaviour
Before it was just a shortcut for install.sh, however that behaviour
completely ignores the fact that install.sh creates multiple files for
configuration on its current directory (or in this case, the directory
of build.sh)

Additionally, trying to change the path for where install.sh should be
breaks the build.sh symlink and symlink check. Since install.sh is being
ran by build.sh, we simply can't tell the user to remove the build.sh

This commit should fix the above issue.
* Add "config" command to install.sh which should do as a workaround to
the install.sh path
* Automatically remove build.sh during "config" install
* Change dir paths when using "config"