[EXP] Per-device configuration
This commit is still in a very early stage, it allows the user to specify configuration for two or more different devices.

This works by replacing the logic of our build command, instead of referencing it from a parameter (build.sh build device-something), get it from a folder. This folder contains various configurations for said device, one device can also contain multiple variation of configurations.

One flaw visible from the early on is the regeneration of the general configuration (located in minibuild/config), every time the user uses regenerate config or when the user creates a new configuration, the general configuration is overwritten. We'll need to come up with a way to avoid this issue.

Current usage:
Call install.sh configexp from build.sh, generate the new configuration setup, then use buildexp to build.

Untested, though by logic it should work fine.